+++ src/NHI/Route.hs
@@ -4,7 +4,6 @@
 
 module NHI.Route where
 
-import Data.Default (Default (..))
 import Data.Map.Strict qualified as Map
 import Data.Maybe (fromJust)
 import Data.Sequence (chunksOf)
@@ -16,6 +15,7 @@ import Ema.Route.Lib.Extra.StaticRoute qualified as SR
 import Ema.Route.Lib.Extra.StringRoute (StringRoute (StringRoute))
 import Ema.Route.Prism (Prism_)
 import Generics.SOP qualified as SOP
+import NHI.PaginatedRoute (Page, PaginatedRoute, getPage)
 import NHI.Types (NixData, Pkg (..))
 import Optics.Core (preview, prism', review)
 
@@ -25,52 +25,6 @@ data Model = Model
   , modelData :: NixData
   }
   deriving stock (Eq, Show, Generic)
-
-newtype Page = Page {unPage :: Int}
-  deriving newtype (Show, Eq, Ord, Num, Enum)
-
-data PaginatedRoute (t :: Type) = PaginatedRoute_Main | PaginatedRoute_OnPage Page
-  deriving stock (Show, Eq, Ord, Generic)
-  deriving anyclass (SOP.Generic, SOP.HasDatatypeInfo)
-
-instance Default (PaginatedRoute t) where
-  def = PaginatedRoute_Main
-
-getPage :: forall {t}. PaginatedRoute t -> Page
-getPage = \case
-  PaginatedRoute_Main -> Page 1
-  PaginatedRoute_OnPage p -> p
-
-fromPage :: forall {t}. Page -> PaginatedRoute t
-fromPage = \case
-  Page 1 -> PaginatedRoute_Main
-  p -> PaginatedRoute_OnPage p
-
-lookupPage :: PaginatedRoute a -> [[a]] -> Maybe [a]
-lookupPage r xs =
-  xs !!? (unPage (getPage r) - 1)
-
-instance IsRoute (PaginatedRoute a) where
-  type RouteModel (PaginatedRoute a) = [[a]]
-  routePrism m =
-    toPrism_ $
-      prism'
-        ( \case
-            PaginatedRoute_Main -> "index.html"
-            PaginatedRoute_OnPage page -> "page/" <> show (unPage page) <> ".html"
-        )
-        ( \fp -> do
-            if fp == "index.html"
-              then pure PaginatedRoute_Main
-              else do
-                page <- fmap toString $ T.stripSuffix ".html" =<< T.stripPrefix "page/" (toText fp)
-                p <- Page <$> readMaybe page
-                void $ lookupPage (fromPage p) m -- Check if this page exists
-                pure $ PaginatedRoute_OnPage p
-        )
-  routeUniverse m =
-    -- TODO: How to tell the other pages to generate?
-    PaginatedRoute_Main : fmap (PaginatedRoute_OnPage . Page) [2 .. (length m)]
 
 type PaginatedListingRoute = PaginatedRoute (Text, NonEmpty Pkg)
